replace Categories and Products components with intefaces
