+++ src/interfaces/CategoryProps.interface.tsx
@@ -1,0 +1,7 @@
+export interface CategoryProps {
+    id: number;
+    name: string;
+    image: string;
+    creationAt: string;
+    updatedAt: string;
+}
